fix: invalid cache data
