+++ internal/parser/parser.go
@@ -1,18 +1,15 @@
 package parser
 
 import (
-	"context"
 	models "github.com/kattana-io/models/pkg/storage"
 	"github.com/kattana-io/tron-blocks-parser/internal/abi"
 	"github.com/kattana-io/tron-blocks-parser/internal/cache"
 	"github.com/kattana-io/tron-blocks-parser/internal/converters"
 	"github.com/kattana-io/tron-blocks-parser/internal/integrations"
-	"github.com/kattana-io/tron-blocks-parser/internal/intermediate"
 	tronApi "github.com/kattana-io/tron-objects-api/pkg/api"
 	"github.com/vmihailenco/msgpack/v5"
 	"go.uber.org/zap"
 	"sync"
-	"time"
 )
 
 type Parser struct {
@@ -103,73 +100,6 @@ func (p *Parser) parseTransaction(transaction tronApi.Transaction) {
 		go p.processLog(log, transaction.TxID, t, owner, &wg)
 	}
 	wg.Wait()
-}
-
-const trxTokenAddress = "TRX"
-const trxDecimals = 6
-
-// GetCachePairToken - Try to pull from cache or populate cache
-func (p *Parser) GetCachePairToken(address string) (string, int32, bool) {
-	pair, err := p.pairsCache.Value(context.Background(), address)
-	if err != nil {
-		pInstance := intermediate.Pair{Address: address}
-		tokenAddress, err := p.api.GetPairToken(address)
-		if err != nil {
-			p.log.Error("GetCachePairToken: " + err.Error())
-			return "", 0, false
-		}
-		// Check list
-		decimals, ok := p.tokenLists.GetDecimals(tokenAddress)
-		if ok {
-			pInstance.SetToken(tokenAddress, decimals)
-		} else {
-			// Call API
-			decimals, err := p.api.GetTokenDecimals(tokenAddress)
-			if err != nil {
-				p.log.Error("GetCachePairToken: GetTokenDecimals: " + err.Error())
-				return "", 0, false
-			}
-			pInstance.SetToken(tokenAddress, decimals)
-		}
-		err = p.pairsCache.Store(context.Background(), address, pInstance, time.Hour*2)
-		if err != nil {
-			p.log.Error("Could not put into cache: " + err.Error())
-		}
-		return pInstance.Token.Address, pInstance.Token.Decimals, true
-	}
-	return pair.Token.Address, pair.Token.Decimals, true
-}
-
-// GetPairTokens - Get tokens of pair
-func (p *Parser) GetPairTokens(address *tronApi.Address) (string, int32, string, int32, bool) {
-	adr, decimals, ok := p.GetCachePairToken(address.ToBase58())
-	if ok {
-		return adr, decimals, trxTokenAddress, trxDecimals, true
-	}
-
-	// Cache miss
-	hexTokenAddress, err := p.api.GetPairToken(address.ToHex())
-	tokenAddr := tronApi.FromHex(hexTokenAddress)
-	cachedDecimals, ok := p.tokenLists.GetDecimals(tokenAddr.ToBase58())
-	if ok {
-		return tokenAddr.ToBase58(), cachedDecimals, trxTokenAddress, trxDecimals, true
-	}
-
-	dec, err := p.api.GetTokenDecimals(tokenAddr.ToHex())
-	if err != nil {
-		p.log.Error("GetPairTokens: " + err.Error())
-		return "", 0, trxTokenAddress, trxDecimals, false
-	}
-	return tokenAddr.ToBase58(), dec, trxTokenAddress, trxDecimals, true
-}
-
-func (p *Parser) GetUniv2PairTokens(address *tronApi.Address) (string, int32, string, int32, bool) {
-	pair, ok := p.jmcache.GetPair(Chain, address)
-	if ok {
-		return pair.TokenA.Address, int32(pair.TokenA.Decimals), pair.TokenB.Address, int32(pair.TokenB.Decimals), true
-	}
-
-	return "", 0, "", 0, false
 }
 
 func (p *Parser) GetEncodedBlock() []byte {
